backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
python-docx
snowflake-connector-python
python-multipart